+++ lib/app.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'homep.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      home: ProfilePage(),
+    );
+  }
+}
